Various improvements to prevent script failing if some applications aren't present.

+++ startup_email.sh
@@ -4,64 +4,89 @@
 # Licensed under GNU Lesser GPL v3
 
 # Originally written for Cygwin and bash
-# Requires smartmontools
+# Some functionality dependent on smartmontools
 
 # TO DO: Rewrite to improve platform support
 
 LOCALDIR=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )
-LOGFILE="${LOCALDIR}/startup_email.log"
-ECONTENTS="${LOCALDIR}/startup_email.contents"
-RECIPIENT="" # Email address for the RECIPIENT
-SUBJECT="[hostname] has started up" # Email subject - change as required
+LOGFILE="${LOCALDIR}/startup_email.log" # Log file
+ECONTENTS="${LOCALDIR}/startup_email.contents" # Temporary file for email contents
+RECIPIENT=$1 # Email address for the recipient (default is command line argument)
+HOSTNAME=`hostname -f` # Get the FQDN (alternatively, enter a hostname)
+SUBJECT=`echo "${HOSTNAME} has started up"` # Email subject - change as required
 ROUTE_TARGET="google.com" # Traceroute target - change as required
 
 date >> ${LOGFILE}
 
+# Email header
+echo '<html>
+<body>' >> ${ECONTENTS}
+
 # Get the external IP address
 echo "Fetching external IP..." >> ${LOGFILE}
-rm -f *.htm*
-wget -q http://checkip.dyndns.org
-echo '<html>
-<body>
-<p style="font-size: 12pt;"><b>External IP: ' >> ${ECONTENTS}
+wget -q http://checkip.dyndns.org -O checkip_results.html
 
-for f in *.htm*
-do
-	host=`tail $f | sed 's/</ /g' | sed 's/>//g' | awk '{print substr($11,1)}'`
-	rm -f $f
-done
-
+# Add the external IP address to the email contents
+echo '<p style="font-size: 12pt;"><b>External IP: ' >> ${ECONTENTS}
+host=`tail checkip_results.html | sed 's/</ /g' | sed 's/>//g' | awk '{print substr($11,1)}'`
 echo $host >> ${ECONTENTS}
 echo $host >> ${LOGFILE}
+rm -f checkip_results.html
 
-#whois $host >> attachment.txt
+#if [ -x "$(command -v whois)" ]; then
+#	whois $host >> attachment.txt
+#fi
 
 # Check the local network interfaces
 echo '</b></p><br /><pre style="font-size: 10pt;"><hr /><b>Network Interfaces:</b><hr />' >> ${ECONTENTS}
-ipconfig /all | grep -E 'Ethernet|Wireless|IP|Gateway' | grep -v 'Description' >> ${ECONTENTS}
+
+if [ -x "$(command -v ipconfig)" ]; then
+	# Windows/Cygwin
+	ipconfig /all | grep -E 'Ethernet|Wireless|IP|Gateway' | grep -v 'Description' >> ${ECONTENTS}
+elif [ -x "$(command -v ifconfig)" ]; then
+	# Linux
+	ifconfig | grep -E 'flags|inet|ether|loop' >> ${ECONTENTS}
+else
+	echo "No tool for querying networ interfaces available." >> ${LOGFILE}
+fi
+
+# Check the route to a nearby network
 echo '<br /><br /><hr /><b>Local Gateway and ISP Routers:</b><hr />' >> ${ECONTENTS}
 
 # Check the path to a nearby network
 echo "Checking network..." >> ${LOGFILE}
-tracert -h 5 -d ${ROUTE_TARGET} | grep -v 'Trace' | grep -v 'hops' | grep -v "^$" >> ${ECONTENTS}
+
+if [ -x "$(command -v tracert)" ]; then
+	# Windows/Cygwin
+	tracert -h 5 -d ${ROUTE_TARGET} | grep -v 'Trace' | grep -v 'hops' | grep -v "^$" >> ${ECONTENTS}
+elif [ -x "$(command -v traceroute)" ]; then
+	# Linux
+	traceroute -m 5 -d ${ROUTE_TARGET} | grep -v 'Trace' | grep -v 'hops' | grep -v "^$" >> ${ECONTENTS}
+else
+	echo "No trace route programme available." >> ${LOGFILE}
+fi
 
 # Check local storage
 echo '<br /><br /><hr /><b>Storage Space:</b><hr />' >> ${ECONTENTS}
 df -h >> ${ECONTENTS}
 
 # Check storage devices
-echo "Scanning storage devices..." >> ${LOGFILE}
-echo '<br /><br /><hr /><b>Storage Device Listing:</b><hr />' >> ${ECONTENTS}
+if [ -x "$(command -v /usr/sbin/smartctl)" ]; then
+	echo "Scanning storage devices..." >> ${LOGFILE}
+	echo '<br /><br /><hr /><b>Storage Device Listing:</b><hr />' >> ${ECONTENTS}
 
-/usr/sbin/smartctl --scan-open >> ${ECONTENTS}
-for d in /dev/sd[a-z]
-do
-	echo "<br /><b>"$d"</b><br />" >> ${ECONTENTS}
-	/usr/sbin/smartctl -i $d | grep -E "Model|Capacity" | awk '{sub(/^/, "\t")};1' >> ${ECONTENTS}
-	echo " " >> ${ECONTENTS}
-	/usr/sbin/smartctl -H $d | grep -v "cygwin" | grep -v "==" | grep -v "Copyright" | grep -v "^$" | awk '{sub(/^/, "\t")};1' >> ${ECONTENTS}
-	/usr/sbin/smartctl -a $d | grep -E "Retired_Block|Reallocated_Sector|Power_|Temperature|Pending|Uncorrectable" | awk '{print "\t" $10 "\t " $2}' | sed 's/_/ /g' | sed 's/Ct/Count/g' >> ${ECONTENTS}
-done
+	/usr/sbin/smartctl --scan-open >> ${ECONTENTS}
+	for d in /dev/sd[a-z]
+	do
+		echo "<br /><b>"$d"</b><br />" >> ${ECONTENTS}
+		/usr/sbin/smartctl -i $d | grep -E "Model|Capacity" | awk '{sub(/^/, "\t")};1' >> ${ECONTENTS}
+		echo " " >> ${ECONTENTS}
+		/usr/sbin/smartctl -H $d | grep -v "cygwin" | grep -v "==" | grep -v "Copyright" | grep -v "^$" | awk '{sub(/^/, "\t")};1' >> ${ECONTENTS}
+		/usr/sbin/smartctl -a $d | grep -E "Retired_Block|Reallocated_Sector|Power_|Temperature|Pending|Uncorrectable" | awk '{print "\t" $10 "\t " $2}' | sed 's/_/ /g' | sed 's/Ct/Count/g' >> ${ECONTENTS}
+	done
+fi
+
+# Footer of email
 echo '</pre></font>
 </body>
 </html>' >> ${ECONTENTS}
@@ -70,10 +95,17 @@ echo '</pre></font>
 echo "Sending email..." >> ${LOGFILE}
 
 # This line could be modified to use a different email script or utility
-python ${LOCALDIR}/mailutils/mail.py -s ${SUBJECT} -r ${RECIPIENT} -l ${ECONTENTS} >> ${LOGFILE}
+python ${LOCALDIR}/mailutils/mail.py -s "${SUBJECT}" -r ${RECIPIENT} -l ${ECONTENTS} >> ${LOGFILE}
+mailresult=$?
 
-# Remove the email contents
-rm -f ${ECONTENTS}
+if [ $mailresult -eq 0 ]; then
+	# Remove the email contents
+	rm -f ${ECONTENTS}
 
-echo "Done.
-" >> ${LOGFILE}
+	echo "Successfully sent email to ${RECIPIENT}" >> ${LOGFILE}
+else
+	echo "Error sending email to  ${RECIPIENT}" >> ${LOGFILE}
+fi
+
+# Add a blank line at the end of the log file entry
+echo >> ${LOGFILE}
